core: presentations: accept a character as '(string 1)
Instead of accepting the character as 'expression i.e. '#\X', we accept the
value '(string 1), i.e 'X'. This is more natural and it is consistent with the
presentation method present that calls princ. Thanks to that it is possible to
read back unparsed command with an argument of the type character.

Improve the presentation method accept specialized on the presentation type
string with regard to defaults: when the string length is specified and
doesn't match then /additionally/ check for the result length being 0 and if
so try to return the default value.