core: define-command: provide better default keyword arguments
The new default value is *unsupplied-argument-marker*. The keyword may remain
unsupplied when it is disabled due to its :when clause returning nil and when
there is no :default option. When :when clause returns t then it is the parser
responsibility to supply a value conforming to the keyword argument ptype.

Since define-command doesn't allow the programmer to specify the variable to
test whether the argument was supplied (like in the ordinary lambda list),
this new default allows them to test with the predicate UNSUPPLIED-ARGUMENT-P.