core: adopt-frame-panes: fix the menu-bar slot in the frame
It is conformant to specify the menu-bar as a list. We have anticipated that
in the function adopt-frame-panes, but later update-frame-panes passes the
value of the menu-bar slot as is and the menu code errors. This commit sets
the slot to the anonymous command table instead to avoid further problems.

Fixes #1237.