core: format-graph-from-roots: don't check for ancestry on reply
Formatting the graph works in two passes: first records are collected and then
they are layed out. After that the graph output record is replayed.

In 4fa59ec4a21b4ab8a439d1c942fc7af6684908af a change was introduced that
skipped replaying the graph if the stream is not drawing and if the graph
record is not a child of the stream's history. The latter was (probably)
introduced as an optimization, but it breaks with-new-output-record, because:

- first a compound record is created (without a parent)
- then the continuation is called (i.e format-graph-from-roots)
- finally the record is added to the stream (or supplied parent)

If the stream were drawing, then we effectively don't see the output. This is
easily reproducible when we wrap f-g-f-r in with-output-as-presentation.

This commit removes the second requriement that the stream's history must be
the ancestor of the graph output record. Fixes #1239.