core: text output: change the initial line height
Previously in both

    STANDARD-EXTENDED-OUTPUT-STREAM
    STANDARD-TEXT-DISPLAYED-OUTPUT-RECORD

we've assumed that the line height starts with 0 and then it is maximized
against added strings. That meant that empty lines had 0 height and putting
couple line breaks was very confusing.

From now on we still maximize the line height but we start from the line
height of the stream's text style.

Fixes #1151.