updating-output-record: remember the line height
Consider the following display function:

   (updating-output (...) (show-a))
   (terpri)
   (updating-output (...) (show-b))

SHOW-A and SHOW-B print a line of text without a line break at the end in a
considerably large font (larger than the default). During initial display we
see two lines of text one below another (thanks to terpri).

The command is executed and it happens that SHOW-B must be called to redisplay
the second output record. To do that an incremental redisplay is invoked:

1. The first record is clean, so set the cursor position at the
   output-record-end-cursor-position
2. Invoke terpri
3. Invoke (show-b)

The problem is with the step 1. because the output-record-end-cursor-position
points to the left-top corner of the cursor without any information about the
cursor height. The cursor height defaults to default text style so terpri
increments the cursor position by the default text style length. Then printing
the line displayed with show-b overlaps with the first "clean" line show-a.

To avoid this problem along with the cursor-x/cursor-y we maintain the cursor
height in the retained graphics state for the record start and end.

Fixes #1240.