incremental-redisplay: change some protocol function signatures
The specification puts the DIFFERENCE-SET in a spilled list of sets. This
commit introduces a backward incompatible change where the difference set is
represented as a list of lists.

Only operators in "21.3 Incremental Redisplay Protocol" are changed. The
function INCREMENTAL-REDISPLAY still accepts spilled list.

The specification is updated to account for these changes.