geometry: introduce a new class STANDARD-REGION-COMPLEMENT
The instance of this class represents a region complement. To compute a
complement we define a generic function REGION-COMPLEMENT. Moreover the
instance of this class may be created as a result of calling REGION-DIFFERENCE
on classes that can't have a simple region representation.

Moreover we remove all uses of the class STANDARD-REGION-DIFFERENCE - all
operations now represented as unions, intersections and complements. This is
the first step towards our geometry being closed under its own operations.

Initially I've considered using STANDARD-REGION-DIFFERENCE where the region A
is +EVERYWHERE+, however REGION-SET is specified to be a subclass of the
bounding rectangle and region complements are unbound.