menu-choose: keyboard navigation of the items
This commit add the possibility to navigate between the items of a menu
with the keyboard. To activate the keyboard navifgation the user
must use MENU-CHOOSE-FROW-DRAWER with a drawer function that
returns three values:
1. a list of navigable presentations,
2. the default presentation (or nil),
3. the keyword :kbd-nav
If the third value isn't :kbd-nav the keyboard navigation isn't
activated and the values returned by the drawer function aren't
used.

If this commit is merged in master is also necessary to document new
features in the McCLIM manual.