recording: use changing-output-record macro
This commit introduces a macro CHANGING-OUTPUT-RECORD that mediates the change
propagation via PROPAGATE-OUTPUT-RECORD-CHANGES. This function is called when
PROPAGATE-OUTPUT-RECORD-CHANGES-P returns true and recomputes changes.

All operations that may affect the output record parent are changed to use
CHANGING-OUTPUT-RECORD instead of calling RECOMPUTE-EXTENT-FOR-*. This allows
us to elegantly inhibit propagating changes while records are changing, i.e

  (defmethod* (setf output-record-position) :around
      (nx ny (record basic-output-record))
    (changing-output-record ((output-record-parent record) record :move)
      (call-next-method)))

    (defmethod propagate-output-record-changes-p
        ((record compound-output-record) child (mode (eql :move)) old-pos old-box)
      (and (not (slot-value record 'in-moving-p))
           (call-next-method)))

Instead of

  (defmethod* (setf output-record-position) :around
      (nx ny (record basic-output-record))
    (with-bounding-rectangle* (min-x min-y max-x max-y) record
      (call-next-method)
      (when-let ((parent (output-record-parent record)))
        (unless (and (typep parent 'compound-output-record)
                     (slot-value parent 'in-moving-p)) ; XXX
          (recompute-extent-for-changed-child parent record
                                              min-x min-y max-x max-y)))
      (values nx ny)))

Default methods:

propagate-output-record-changes-p - parent is not null and the bounding
rectangle changes

propagate-output-record-changes - depending on the mode it calls recompute-*
or nothing; additionally it returns the difference-set and check-overlapping
values unchanged.