core: transform-coordinates-mixin: ensure the identity transformation
TRANSFORM-COORDINATES-MIXIN is responsible for pretransforming in the :AROUND
method coordinates of drawing operations with the MEDIUM-TRANSFORMATION.

From now on it wraps the next method in WITH-IDENTITY-TRANSFORMATION after
transforming coordinates so the downstream method doesn't see it. This is
important for fallback methods and removes the necessity for removing the
transformation and untransforming coordinates in them.

Also it is less surprising - the downstream method may safely try to transform
its coordinates with the medium transformation and that operation will be
idempotent, so for the backend methods it doesn't really matter whether
TRANSFORM-COORDINATES-MIXIN is the medium superclass or not.