core: with-output-to-pixmap: trampoline to a generic function
Previously the macro expanded directly to a code allocating a pixmap. From now
on it expands to a call to the function INVOKE-WITH-OUTPUT-TO-PIXMAP. Two
methods are provided:

- sheet specialization that calls the method with its medium bound

- basic-medium specialization that:

  1. tries to fix missing :width / :height args when possible
  2. allocates a new pixmap and makes it a drawable of a new medium
  3. clears the pixmap with the medium background ink
  4. calls the continuation and returns the pixmap

When either width or height is missing and the medium sheet is the output
recording stream, then we call the continuation to create the output record -
this is used only to set the width and/or the height. The continuation is
invoked once more in the step (4.) above.