core: design-ink: return only the topmost design (shallow lookup)
This is important because designs are over-composed and the transformation
appleis only to its own layer.

Previously

(design-ink design x y)
 -> (design-ink inner-at-transformed-xy transformed-x transformed-y)

Now

(design-ink design x y)
 -> inner-at-transformed-xy

So when we want to learn about the rgba value at a specified coordiante we may
now apply untransformed x, y with:

(design-ink inner-at-transformed-xy x y)

And that gives the correct result. A function that descends recursively but
retains the correct corrsepondence between layer is introduced: DESIGN-INK*.