core: with-output-to-drawing-stream: use WITH-PORT
This approach ensures that when the port was created fro this particular
invocation then it is destroyed afterwards. It is still possible to retain the
port (per WITH-PORT behavior), by supplying a port that already exists to the
macro.

Additionally make the NULL specialization a little more robust - when the
destination is NIL and there is no continuation nor backend specialization, then
we use FIND-PORT - we don't want to destroy the interactive port after the
window is created, because that would also destroy the window. We also specify
the window input buffer to nil to avoid race conditions with the event queue.