core: transform-coordinates-mixin: handle line-unit :coordinates
Previously we've fixed the medium transformation to identity so coordinates
won't be transformed twice - this affected a situation when the line unit is
:COORDINATES, because LINE-STYLE-EFFECTIVE-{THICKNESS,DASHES} relies on the
medium transformation.

We replace the invocation of WITH-IDENTITY-TRANSFORMATION with a macro that
dynamically sets the line style with the unit :normal and the thickness and
dashes containing precomputed effective values.

Additionally:

- make effective-line-style-{thickness,dashes} work for ungrafted mediums when
  the unit is :point (the unit defaults then to the value of :normal)

- add a dwim macro WITH-TRANSFORMED-POSITIONS* that accepts either a variable
  with a sequence containing positions or a sequence of variables to transform

- make WITH-IDENTITY-TRANSFORMATION* (additionally to supplementing the line)
  use WITH-TRANSFORMED-POSITIONS* (that results in much smaller code)