core: sheets: more async-friendly implementation of sheet readers
SHEET-{NATIVE,DEVICE}-{TRANSFORMATION,REGION} were constructed like this:

(with-slots (thing)
  (unless thing
    (setf thing (the-thing)))
  thing)

This could be troublesome when something asynchronously i.e resized the sheet,
because between setting and returning the slot, it could be invalidated. To
avoid this type of issue we do:

(or (slot-value sheet 'thing)
    (setf (slot-value sheet 'thing) (the-thing)))

And this is guaranteed to always return something reasonable.