extensions: render: optimize functions fill-image{-with-mask}
This optimization is tailored for easy cases like a solid color, but also for
drawing with stencil.

- numerous declarations
- multiple expansion of the same code inside typecase
- with-brush macro that provides necessary operators at compile time
- add an utility DO-REGIONS* that does not bind the second variable

This optimization yields 5x improvement for a solid color, 4x improvement for
flipping ink etc. This commit improves only the performance for rectangles and
text rendering (the latter much less than for rectangles).