core: with-output-as-gadget: create a trampoline
The macro WITH-OUTPUT-AS-GADGET was a "naked" macro and not a trampoline to
the function INVOKE-WITH-OUTPUT-AS-GADGET - that was against coding
conventions specified for CLIM. This commit introduces the latter.

Also remove an obsolete comment about calling the gadget constructor on each
redisplay in incremental redisplay to prevent an error. That problem was fixed
in the commit 6c1c0ca9bc3eeb075520323d302aec747f139c8a, see issue #1247.