core: %accept: ensure congruent lambda lists
Some presentation methods for ACCEPT does not mention parameters DEFAULT and
DEFAULT-TYPE - this is fine, but if we take Common Lisp standard to the letter
then method must mention all keyword arguments in the generic function. One of
such meticulous implementations is Mezzano. To avoid errors do the same thing we
did for PRESENT.