core: repaint: improve the protocol methods
- previously repaint-sheet had two auxilliary :AFTER methods for propagating
repaint to children and for finishing output. Depending on the order of the
sheet superclasses SHEET-PARENT-MIXIN and SHEET-WITH-MEDIUM-MIXIN the order of
evaluation of :AFTER methods was different:

1. handle-repaint, propagate-repaint, and finish-output, or
2. handle-repaint, finish-output, and propagate-repaint

the second scenario may lead to an error, when the method medium-finish-output
ignores mediums without a sheet with the direct mirror and flushes only the
mirrored medium. In that case we sap the buffer with cleaned background and
then we repaint children - no output is visible

- we are more precise with computing regions for the repaint and for clearing
the medium area by introducing a helper function SHEET-VISIBLE-AREA.  See the
issue #1280 to learn details behind the function SHEET-NATIVE-REGION*.

- REPAINT-SHEET calls REPAINT-SHEET recursively on its children. This is more
compatible with the spec; previously we've skipped mirrors, but that was not
correct because REPAINT-SHEET may be also called from the user code, and that
may be i.e because the model behind the child mirrored sheet also changed

- REPAINT-SHEET on SHEET-MUTE-REPAINTING-MIXIN now does nothing. This is what
the specification says (while we've defined empty method HANDLE-REPAINT)