backends: clx: fix a typo in port-set-mirror-geometry
A silly mistake where the bounding-rectangle* result was bound as

  (x1 x2 y1 y2)

instead of

  (x1 y1 x2 y2)

that resulted in invalid recomputation of the size of a horizontal scroll
bar (and possibly other things too).

That said this was invalid nevertheless - we need to compare dimensions, not
the second coordinate (dimension may change even when the second coordinate is
the same, i.e when the beginning is shrunk).