core: layout protocol: make merge-one-option recognize :compute
When a space requirement component is :COMPUTE, then the value returned by
COMPOSE-SPACE is used. This allows us to remove a hack that required us to
overwrite user options in the method for COMPOSE-SPACE :AROUND on
CLIM-STREAM-PANE.

We don't redisplay the history in COMPOSE-SPACE - we take it as is. When
REDISPLAY-FRAME-PANE changes the size requirements, then the layout is
recomputed either way.

Also remove accessors that are not necessary anymore.