core: changing-space-requirements: saner implementation of batching
- previously CHANGING-SPACE-REQUIREMENTS inhibited only CSR invoked for the
TOP-LEVEL-SHEET-PANE, so when the invocation was not passed all the way up,
CHANGE-SPACE-REQUIREMENTS was eagerly resolved, i.e ALLOCATE-SPACE was called.

- instead of calling LAYOUT-FRAME indiscriminately, we call
NOTE-SPACE-REQUIREMENTS-CHANGED for panes that changed requirements and have
not without supplied RESIZE-FRAME argument to CHANGE-SPACE-REQUIREMENTS.

- avoid unnecessary recomputations by merging requirement changes - when a
frame needs to have a new layout, then none of its panes need a separate call
to NOTE-SPACE-REQUIREMENTS-CHANGED.

- invalidate cache space requirements of pane and its parents (this may change
in the future)

- provide a meaningful interpretation for the argument LAYOUT in
CHANGING-SPACE-REQUIREMENTS for a pane:

cond    RESIZE-SHEET (layout-frame frame)
        LAYOUT       (layout-frame frame current-width current-height)
        otherwise    (note-space-requirements-changed parent pane)

- additionally: INVOKE-WITH-CHANGING-SPACE-REQUIREMENTS mistakenly tested
for *CHANGED-SPACE-REQUIREMENTS* and not for *CHANGING-SPACE-REQUIREMENTS*, so
some calls were not batched (and did not trigger the layout protocol).