core: event compression: prepend configuration events
The event compression of repaint and configuration events may cause an issue
where resize event is reordered after its corresponding repaint event. For
example consider the following events:

(in this example the initial size is +nowhere+)

1. resize   0   0 100 100
2. expose   0   0 100 100
3. resize   0   0 150 150
4. expose 100 100  50  50
5. resize 200 200 350 350 (this does not change window coordinates)

So the last configuration event comes after the last repaint event. After
compression we have:

1. expose 0 0 150 150
2. resize 0 0 150 150

First we handle the "expose", and repaint nothing, because the intersection of
the sheet region and the repaint region is nowhere. Then we handle the
"resize" that changes the window geometry. The effect is that we have the
window with the correcte size but with no content.

"resize" is triggered by the window-configuration-event
"expose" is triggered by the window-repaint-event