core: w-r-f-g: write an implementation consistent between methods
Previously WITH-ROOM-FOR-GRAPHICS made various adjustements to the created
record, and that led to inconsistencies between "vanilla" and "recorded"
implementations. Now we do not clip the graphic beyond the current clipping
region, and adjust the cursor according to parameters WIDTH and HEIGHT.

The origin is set to the baseline when FIRST-QUADRANT is T, and to the cursor
position otherwise.