core: configuration event: call dispatch-repaint from handle-event
Previously we've called dispatch-repaint when the sheet region was changed,
however this is too naive - we either repaint only damanged parts, or whole
sheet and there is no telling, what is the desired behavior.

Instead of that we expect now, that repaint sheet will be invoked by the
caller. This allows for better control in handle-event specializations. This
commit does not introduce any changes in behavior of "normal" panes.