backends: clx: don't set up the background and remove a hack
When CLX backend created a window, it also set its background. That lead to a
significant flicker, because on resize the server first repainted the
background, and after that send the window-repaint-event (that could be handled
with some delay because of being in the event queue). Then we've repaint the
background again and then painted the actual content.

Moreover we remove the bit-gravity "evil" hack that set it to :north-west for
EXTENDED-OUTPUT-STREAMs. This is counterproductive if we want to support output
records that change their replay method depending on the sheet size. Also this
hack lead to a non-consistent behavior between sheets.