present: decouple arguments ALLOW-SENSITIVE-INFERIORS and SENSITIVE
Preivously when SENSITIVE was NIL, then neither presentation nor its inferiors
were sensitive (disregarding of the value of ALLOW-SENSITIVE-INFERIORS).

The new interpretation is to make SENSITIVE apply only to the current call to
PRESENT so it is possible to make the presentation not sensitive but keep its
inferiors sensitive when ALLOW-SENSITIVE-INFERIORS is true.

To retain partial backward compatibility ALLOW-SENSITIVE-INFERIORS defaults to
the value of SENSITIVE so calls like

    (present foo type :sensitive nil)

will behave the same as earlier - neither the presentation nor its inferiors are
sensitive. However it is possible to change this behavior with

    (present foo type :sensitive nil :allow-sensitive-inferiors t)

and then all inferiors will be sensitive while the object foo won't. This new
behavior is useful when we present an object in a different view and we don't
want for it to be selectable but we have certain knobs presented along with it
that may be used to configure said object.