Wayland FFI: Rethink REALIZE-MIRROR a little
Previously we were creating a mirror when the graft was initialized and then
assigning this same mirror to the top level sheet.

I'm now not fully convinced that a graft needs to have a mirror that
represents a window in the native server. Instead that mirror should live more
directly with the TOP-LEVEL-SHEETs

In this we're moving logic that is specific to the top level
surface (e.g. window creation and mapping) to be part of the
TOP-LEVEL-SHEET-MIXIN specialization instead of MIRRORED-SHEET-MIXIN.